DeatNote renamed to KillTome, deleted DeathNoteSpawn event
