--- Content.Shared/DeathNote/DeathNoteComponent.cs
@@ -1,11 +1,0 @@
-using Robust.Shared.GameStates;
-
-namespace Content.Shared.DeathNote;
-
-/// <summary>
-/// Paper with that component is Death Note.
-/// </summary>
-[RegisterComponent, NetworkedComponent]
-public sealed partial class DeathNoteComponent : Component
-{
-}
