+++ Content.Shared/KillTome/KillTomeComponent.cs
@@ -1,0 +1,11 @@
+using Robust.Shared.GameStates;
+
+namespace Content.Shared.KillTome;
+
+/// <summary>
+/// Paper with that component is KillTome.
+/// </summary>
+[RegisterComponent, NetworkedComponent]
+public sealed partial class KillTomeComponent : Component
+{
+}
